v0.69 PR4 (Phase 3 complete) — split remaining 5 ai-edit.ts sections
Completes Plan G Phase 3. With PR #166 (3/8) + this PR (5/8), all 8
ai-edit.ts sections now live in `commands/_shared/edit/`. ai-edit.ts
collapses from 1,589 → 79 lines (−95%) — a pure re-export barrel.

## Splits in this PR (5)

| File | Lines | Source |
|---|---:|---|
| `_shared/edit/silence-cut.ts` | 347 | FFmpeg silencedetect + Gemini fallback |
| `_shared/edit/jump-cut.ts` | 297 | Whisper word-level + filler removal |
| `_shared/edit/caption.ts` | 283 | Whisper + ASS subtitles burn-in (Remotion fallback) |
| `_shared/edit/translate-srt.ts` | 190 | Claude/OpenAI SRT translation, batched |
| `_shared/edit/text-overlay.ts` | 287 | drawtext filter (system font auto-detect, Remotion fallback) |
| `_shared/edit/_helpers.ts` | 31 | private getVideoResolution + escapeDrawtext |

## ai-edit.ts: barrel only

79 lines, all re-exports. The 9 consumer files (demo, edit-cmd,
ai-edit-cli, ai-review, ai-script-pipeline-cli, ai-animated-caption,
ai, ai-script-pipeline, ai.test) work unchanged.

## Phase 3 totals (PR #166 + this PR)

ai-edit.ts: 1,589 → 79 lines (−1,510 source lines).
9 new files under `commands/_shared/edit/`: 1,711 total lines (split + 1
helpers file). Net +201 lines (mostly module-doc headers + import
boilerplate per file).

## Test plan

- [x] `pnpm -r exec tsc --noEmit` — 0 errors
- [x] `pnpm lint` — 0 errors
- [x] `pnpm -F @vibeframe/cli test` — 652 passed (unchanged)
- [x] All re-exports preserve external import paths
